+++ apps/web/lingodotdev/server.ts
@@ -2,8 +2,8 @@ import { createInstance } from "i18next";
 import ICU from "i18next-icu";
 import resourcesToBackend from "i18next-resources-to-backend";
 import { initReactI18next } from "react-i18next/initReactI18next";
+import { DEFAULT_LOCALE } from "@/lib/constants";
 import { getLocale } from "@/lingodotdev/language";
-import { DEFAULT_LANGUAGE } from "@/lingodotdev/shared";
 
 const initI18next = async (lng: string) => {
   const i18nInstance = createInstance();
@@ -13,7 +13,7 @@ const initI18next = async (lng: string) => {
     .use(resourcesToBackend((language: string) => import(`../locales/${language}.json`)))
     .init({
       lng,
-      fallbackLng: DEFAULT_LANGUAGE,
+      fallbackLng: DEFAULT_LOCALE,
       interpolation: {
         escapeValue: false,
       },
