--- apps/web/lingodotdev/shared.ts
@@ -1,1 +1,0 @@
-export const DEFAULT_LANGUAGE = "en-US";
